added Tophat2, STAR and Bowtie2(L) aln (PE & SE)
did not thoroughly test bowtie2(L) script
need to change DetectAllelicConcatenated for STAR to MAPQ == 255 (replace MAPQ >= 1, which will work fine for Tophat2 and Bowtie2).
tophat2 has the bug where the indexed genome is not within the folder containing the fasta file (it is in a subdirectory called /bowtie2-index/). Bowtie2 automatically reconstructs the fasta file from the index, which enables the pipeline to continue but this is not optimal.